move table files
