--- src/features/table/fileCard/contextMenu/DownloadMenuItem.tsx
@@ -1,58 +1,0 @@
-import { ContextMenuItem } from "@src/components/ui/context-menu"
-import { useToast } from "@src/components/ui/use-toast"
-import { saveFolder } from "@src/features/fileDownload"
-import { generateQuickLinkInfo } from "@src/features/quickAccess"
-import { downloadFile } from "@src/lib/utils"
-import { DownloadIcon } from "lucide-react"
-
-interface DownloadMenuItemProps {
-    href: string
-    isFolder: boolean
-    fileName: string
-}
-
-export default function DownloadMenuItem({
-    href,
-    isFolder,
-    fileName,
-}: DownloadMenuItemProps) {
-    const { toast } = useToast()
-
-    const handleFolderDownload = async () => {
-        const { name } = generateQuickLinkInfo(href)
-        toast({
-            title: "📥 Downloading",
-            description: "Fetching and archiving files.",
-        })
-        try {
-            await saveFolder(href)
-            toast({
-                title: "✅ Success",
-                description: `Downloaded ${name}.`,
-            })
-        } catch (error) {
-            toast({
-                title: "❌ Error",
-                description: `Failed to download ${name}.`,
-            })
-        }
-    }
-
-    if (isFolder) {
-        return (
-            <ContextMenuItem onSelect={handleFolderDownload}>
-                <DownloadIcon /> Download folder
-            </ContextMenuItem>
-        )
-    }
-
-    const handleSelect = () => {
-        downloadFile(href, fileName)
-    }
-
-    return (
-        <ContextMenuItem onSelect={handleSelect}>
-            <DownloadIcon /> Download file
-        </ContextMenuItem>
-    )
-}
