+++ src/features/table/structure/TableSkeleton.tsx
@@ -1,4 +1,4 @@
-import { Skeleton } from "../../components/ui/skeleton"
+import { Skeleton } from "../../../components/ui/skeleton"
 
 export default function TableSkeleton() {
     return (
